+++ src/api/iscomputed.ts
@@ -1,13 +1,13 @@
 import { isObservableObject } from "../types/observableobject"
 import { getAtom } from "../types/type-utils"
 import { isComputedValue } from "../core/computedvalue"
-import { fail } from "../utils/utils"
+import { fail, $mobx } from "../utils/utils"
 
 export function _isComputed(value, property?: string): boolean {
     if (value === null || value === undefined) return false
     if (property !== undefined) {
         if (isObservableObject(value) === false) return false
-        if (!value.$mobx.values[property]) return false
+        if (!value[$mobx].values[property]) return false
         const atom = getAtom(value, property)
         return isComputedValue(atom)
     }