+++ src/api/isobservable.ts
@@ -4,7 +4,7 @@ import { isObservableObject, ObservableObjectAdministration } from "../types/obs
 import { isAtom } from "../core/atom"
 import { isComputedValue } from "../core/computedvalue"
 import { isReaction } from "../core/reaction"
-import { fail } from "../utils/utils"
+import { fail, $mobx } from "../utils/utils"
 
 function _isObservable(value, property?: string): boolean {
     if (value === null || value === undefined) return false
@@ -17,7 +17,7 @@ function _isObservable(value, property?: string): boolean {
                 "isObservable(object, propertyName) is not supported for arrays and maps. Use map.has or array.length instead."
             )
         if (isObservableObject(value)) {
-            const o = <ObservableObjectAdministration>(value as any).$mobx
+            const o = <ObservableObjectAdministration>(value as any)[$mobx]
             return o.values && !!o.values[property]
         }
         return false
@@ -25,7 +25,7 @@ function _isObservable(value, property?: string): boolean {
     // For first check, see #701
     return (
         isObservableObject(value) ||
-        !!value.$mobx ||
+        !!value[$mobx] ||
         isAtom(value) ||
         isReaction(value) ||
         isComputedValue(value)