+++ src/api/when.ts
@@ -1,4 +1,4 @@
-import { Lambda, fail, getNextId } from "../utils/utils"
+import { Lambda, fail, getNextId, $mobx } from "../utils/utils"
 import { IReactionDisposer } from "../core/reaction"
 import { autorun } from "./autorun"
 import { createAction } from "../core/action"
@@ -28,7 +28,7 @@ function _when(predicate: () => boolean, effect: Lambda, opts: IWhenOptions): IR
     let timeoutHandle: any
     if (typeof opts.timeout === "number") {
         timeoutHandle = setTimeout(() => {
-            if (!disposer.$mobx.isDisposed) {
+            if (!disposer[$mobx].isDisposed) {
                 disposer()
                 const error = new Error("WHEN_TIMEOUT")
                 if (opts.onError) opts.onError(error)