+++ src/types/type-utils.ts
@@ -1,5 +1,5 @@
 import { IDepTreeNode } from "../core/observable"
-import { fail } from "../utils/utils"
+import { fail, $mobx } from "../utils/utils"
 import { initializeInstance } from "../utils/decorators2"
 import { isAtom } from "../core/atom"
 import { isComputedValue } from "../core/computedvalue"
@@ -16,7 +16,7 @@ export function getAtom(thing: any, property?: string): IDepTreeNode {
                     process.env.NODE_ENV !== "production" &&
                         "It is not possible to get index atoms from arrays"
                 )
-            return (thing as any).$mobx.atom
+            return (thing as any)[$mobx].atom
         }
         if (isObservableMap(thing)) {
             const anyThing = thing as any
@@ -33,11 +33,11 @@ export function getAtom(thing: any, property?: string): IDepTreeNode {
         }
         // Initializers run lazily when transpiling to babel, so make sure they are run...
         initializeInstance(thing)
-        if (property && !thing.$mobx) thing[property] // See #1072
+        if (property && !thing[$mobx]) thing[property] // See #1072
         if (isObservableObject(thing)) {
             if (!property)
                 return fail(process.env.NODE_ENV !== "production" && `please specify a property`)
-            const observable = (thing as any).$mobx.values[property]
+            const observable = (thing as any)[$mobx].values[property]
             if (!observable)
                 fail(
                     process.env.NODE_ENV !== "production" &&
@@ -51,9 +51,9 @@ export function getAtom(thing: any, property?: string): IDepTreeNode {
             return thing
         }
     } else if (typeof thing === "function") {
-        if (isReaction(thing.$mobx)) {
+        if (isReaction(thing[$mobx])) {
             // disposer function
-            return thing.$mobx
+            return thing[$mobx]
         }
     }
     return fail(process.env.NODE_ENV !== "production" && "Cannot obtain atom from " + thing)
@@ -66,7 +66,7 @@ export function getAdministration(thing: any, property?: string) {
     if (isObservableMap(thing)) return thing
     // Initializers run lazily when transpiling to babel, so make sure they are run...
     initializeInstance(thing)
-    if (thing.$mobx) return thing.$mobx // TODO: use a Symbol
+    if (thing[$mobx]) return thing[$mobx] // TODO: use a Symbol
     fail(process.env.NODE_ENV !== "production" && "Cannot obtain administration from " + thing)
 }